+++ examples/arduino/Phone/ESP_Panel_Board_Custom.h
@@ -25,6 +25,7 @@
  *      - EK9716B
  *      - GC9A01, GC9B71, GC9503
  *      - ILI9341
+ *      - JD9365
  *      - NV3022B
  *      - SH8601
  *      - SPD2010
@@ -110,8 +111,6 @@
                                                         // |--------------|---------------|
     #define ESP_PANEL_LCD_RGB_DATA_WIDTH        (16)    // |      8       |      16       |
     #define ESP_PANEL_LCD_RGB_PIXEL_BITS        (16)    // |      24      |      16       |
-
-    #define ESP_PANEL_LCD_RGB_FRAME_BUF_NUM     (1)     // 1/2/3
     #define ESP_PANEL_LCD_RGB_BOUNCE_BUF_SIZE   (0)     // Bounce buffer size in bytes. This function is used to avoid screen drift.
                                                         // To enable the bounce buffer, set it to a non-zero value. Typically set to `ESP_PANEL_LCD_WIDTH * 10`
                                                         // The size of the Bounce Buffer must satisfy `width_of_lcd * height_of_lcd = size_of_buffer * N`,
@@ -121,7 +120,6 @@
     #define ESP_PANEL_LCD_RGB_IO_DE             (17)    // -1 if not used
     #define ESP_PANEL_LCD_RGB_IO_PCLK           (9)
     #define ESP_PANEL_LCD_RGB_IO_DISP           (-1)    // -1 if not used
-
                                                         // | RGB565 | RGB666 | RGB888 |
                                                         // |--------|--------|--------|
     #define ESP_PANEL_LCD_RGB_IO_DATA0          (10)    // |   B0   |  B0-1  |   B0-3 |
@@ -158,6 +156,22 @@
                                                                 // The `mirror()` function will be implemented by LCD command if set to 1.
 #endif
 
+#elif ESP_PANEL_LCD_BUS_TYPE == ESP_PANEL_BUS_TYPE_MIPI_DSI
+
+    #define ESP_PANEL_LCD_MIPI_DSI_LANE_NUM         (2)     // ESP32-P4 supports 1 or 2 lanes
+    #define ESP_PANEL_LCD_MIPI_DSI_LANE_RATE_MBPS   (1000)  // Single lane bit rate, should consult the LCD supplier or check the
+                                                            // LCD drive IC datasheet for the supported lane rate.
+                                                            // ESP32-P4 supports max 1500Mbps
+    #define ESP_PANEL_LCD_MIPI_DSI_PHY_LDO_ID       (3)     // -1 if not used
+    #define ESP_PANEL_LCD_MIPI_DPI_CLK_MHZ          (52)
+    #define ESP_PANEL_LCD_MIPI_DPI_PIXEL_BITS       (ESP_PANEL_LCD_RGB565_COLOR_BITS_16)
+    #define ESP_PANEL_LCD_MIPI_DSI_HPW              (10)
+    #define ESP_PANEL_LCD_MIPI_DSI_HBP              (160)
+    #define ESP_PANEL_LCD_MIPI_DSI_HFP              (160)
+    #define ESP_PANEL_LCD_MIPI_DSI_VPW              (1)
+    #define ESP_PANEL_LCD_MIPI_DSI_VBP              (23)
+    #define ESP_PANEL_LCD_MIPI_DSI_VFP              (12)
+
 #else
 
 #error "The function is not ready and will be implemented in the future."
@@ -176,18 +190,20 @@
  *   2. Formatter: ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(delay_ms, command, { data0, data1, ... }) and
  *                ESP_PANEL_LCD_CMD_WITH_NONE_PARAM(delay_ms, command)
  */
-// #define ESP_PANEL_LCD_VENDOR_INIT_CMD()                                        \
-//     {                                                                          \
-//         {0xFF, (uint8_t []){0x77, 0x01, 0x00, 0x00, 0x10}, 5, 0},              \
-//         {0xC0, (uint8_t []){0x3B, 0x00}, 2, 0},                                \
-//         {0xC1, (uint8_t []){0x0D, 0x02}, 2, 0},                                \
-//         {0x29, (uint8_t []){0x00}, 0, 120},                                    \
-//         or                                                                     \
-//         ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xFF, {0x77, 0x01, 0x00, 0x00, 0x10}), \
-//         ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC0, {0x3B, 0x00}),                   \
-//         ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC1, {0x0D, 0x02}),                   \
-//         ESP_PANEL_LCD_CMD_WITH_NONE_PARAM(120, 0x29),                               \
-//     }
+/*
+#define ESP_PANEL_LCD_VENDOR_INIT_CMD()                                        \
+    {                                                                          \
+        {0xFF, (uint8_t []){0x77, 0x01, 0x00, 0x00, 0x10}, 5, 0},              \
+        {0xC0, (uint8_t []){0x3B, 0x00}, 2, 0},                                \
+        {0xC1, (uint8_t []){0x0D, 0x02}, 2, 0},                                \
+        {0x29, (uint8_t []){0x00}, 0, 120},                                    \
+        or                                                                     \
+        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xFF, {0x77, 0x01, 0x00, 0x00, 0x10}), \
+        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC0, {0x3B, 0x00}),                   \
+        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC1, {0x0D, 0x02}),                   \
+        ESP_PANEL_LCD_CMD_WITH_NONE_PARAM(120, 0x29),                               \
+    }
+*/
 
 /* LCD Color Settings */
 /* LCD color depth in bits */
@@ -378,8 +394,8 @@
  *
  */
 #define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MAJOR 0
-#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MINOR 2
-#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_PATCH 2
+#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MINOR 3
+#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_PATCH 1
 
 #endif /* ESP_PANEL_USE_CUSTOM_BOARD */