+++ src/systems/phone/stylesheets/default/dark/recents_screen_data.h
@@ -56,7 +56,7 @@ extern "C" {
         .trash_icon = {                                                           \
             .default_size = ESP_BROOKESIA_STYLE_SIZE_SQUARE(48),                         \
             .press_size = ESP_BROOKESIA_STYLE_SIZE_SQUARE(43),                           \
-            .image = ESP_BROOKESIA_STYLE_IMAGE(&esp_brookesia_image_large_recents_screen_trash_48_48),\
+            .image = ESP_BROOKESIA_STYLE_IMAGE(&esp_brookesia_image_large_recents_screen_trash_64_64),\
         },                                                                        \
         .flags = {                                                                \
             .enable_memory = 1,                                                   \