+++ src/systems/phone/stylesheets/default/dark/stylesheet.h
@@ -58,6 +58,7 @@ extern "C" {
         },                                                   \
         .flags = {                                           \
             .enable_gesture = 1,                             \
+            .enable_gesture_navigation_back = 0,             \
             .enable_recents_screen_snapshot_drag = 1,        \
             .enable_recents_screen_hide_when_no_snapshot = 1,                     \
         },                                                   \