+++ src/systems/phone/stylesheets/stylesheets.h
@@ -10,6 +10,7 @@
 #include "320_240/dark/stylesheet.h"
 #include "320_480/dark/stylesheet.h"
 #include "480_480/dark/stylesheet.h"
+#include "720_1280/dark/stylesheet.h"
 #include "800_480/dark/stylesheet.h"
 #include "800_1280/dark/stylesheet.h"
 #include "1024_600/dark/stylesheet.h"